Fix FTBFS due to -Werror=alloc-size-larger-than on GCC 12
GCC 12 emits an error when compiling squid due to
-Werror=alloc-size-larger-than.  The error looks like this:

store/Disks.cc:690:64: error: argument 1 value '18446744073709551615' exceeds maximum object size 9223372036854775807 [-Werror=alloc-size-larger-than=]
  690 |         const auto tmp = new SwapDir::Pointer[swap->n_allocated];
      |                                                                ^

Arguably, this is likely a compiler bug and not squid's fault.  Either
way, I believe it's worth having more type safety in this code.

+++ src/store/Disks.cc
@@ -787,7 +787,7 @@ allocate_new_swapdir(Store::DiskConfig &swap)
         swap.swapDirs = new SwapDir::Pointer[swap.n_allocated];
     }
 
-    if (swap.n_allocated == swap.n_configured) {
+    if (swap.n_allocated == (unsigned int) swap.n_configured) {
         swap.n_allocated <<= 1;
         const auto tmp = new SwapDir::Pointer[swap.n_allocated];
         for (int i = 0; i < swap.n_configured; ++i) {
